Update elli_lmug
Rework handle/2 such that the second argument is a list of the form:

    '[#(lmug-mw-mod-name [#(mw-opt value) ...]
      ...]

Currently, if any element of `handlers` (the list above) is
poorly formed or if the specified module does not export wrap/2, i.e.
does not implement the lmug-mw behaviour, then handle/2 returns 'ignore
and stops processing the rest of the handlers.

do-handle/3 has been renamed handle/3 and performs the iteration.